refactor: move ai implementation sagas
- one file per saga
- subfolder for sagas
