+++ src/data/aiImplementations/aiImplementationSagas/fetchAiImplementationHealth.ts
@@ -1,0 +1,36 @@
+import { put } from 'redux-saga/effects'
+
+import urlBuilder from '../../util/urlBuilder'
+import httpResponseErrorMessage from '../../util/httpResponseErrorMessage'
+import { aiImplementationHealthDataActions } from '../aiImplementationListActions'
+
+// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
+export default function* fetchAiImplementationHealth(
+  aiImplementationId: string,
+) {
+  try {
+    const response = yield fetch(
+      urlBuilder(`ai-implementations/${aiImplementationId}/health-check`),
+      {
+        method: 'GET',
+      },
+    )
+
+    if (!response.ok) {
+      throw new Error(httpResponseErrorMessage(response))
+    }
+
+    const data = yield response.json()
+
+    yield put(
+      aiImplementationHealthDataActions.store(data.status, aiImplementationId),
+    )
+  } catch (error) {
+    yield put(
+      aiImplementationHealthDataActions.errored(
+        `Failed to fetch AI implementation list: ${error.message}`,
+        aiImplementationId,
+      ),
+    )
+  }
+}
