+++ src/data/rootSaga.ts
@@ -1,11 +1,11 @@
 import { all } from 'redux-saga/effects'
 import caseSetSagas from './caseSets/caseSetSagas'
-import aiImplementationListSagas from './aiImplementations/aiImplementationListSagas'
+import aiImplementationSagas from './aiImplementations/aiImplementationSagas'
 import benchmarkSagas from './benchmarks/benchmarkSagas'
 
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 export default function* rootSaga() {
   yield all([...caseSetSagas])
-  yield all([...aiImplementationListSagas])
+  yield all([...aiImplementationSagas])
   yield all([...benchmarkSagas])
 }
